+++ ext/src/linq/src/api/dbfunc/DbFunc.DateTime.cs
@@ -119,143 +119,16 @@ namespace mooSQL.linq
 
 		#region DateDiff
 
-		sealed class DateDiffBuilderSapHana : IExtensionCallBuilder
-		{
-			public void Build(ISqExtensionBuilder builder)
-			{
-				var part       = builder.GetValue<DateParts>(0);
-				var startdate  = builder.GetExpression(1);
-				var endDate    = builder.GetExpression(2);
-				var divider    = 1;
-
-				if (startdate is null || endDate is null)
-				{
-					builder.IsConvertible = false;
-					return;
-				}
-
-				string funcName;
-				switch (part)
-				{
-					case DateParts.Day        : funcName = "Days_Between";                     break;
-					case DateParts.Hour       : funcName = "Seconds_Between"; divider = 3600;  break;
-					case DateParts.Minute     : funcName = "Seconds_Between"; divider = 60;    break;
-					case DateParts.Second     : funcName = "Seconds_Between";                  break;
-					case DateParts.Millisecond: funcName = "Nano100_Between"; divider = 10000; break;
-					default:
-						throw new InvalidOperationException($"Unexpected datepart: {part}");
-				}
-
-				IExpWord func = new FunctionWord(typeof(int), funcName, startdate, endDate);
-				if (divider != 1)
-					func = builder.Div(func, divider);
-
-				builder.ResultExpression = func;
-			}
-		}
-
-		sealed class DateDiffBuilderDB2 : IExtensionCallBuilder
-		{
-			public void Build(ISqExtensionBuilder builder)
-			{
-				var part       = builder.GetValue<DateParts>(0);
-				var startDate  = builder.GetExpression(1);
-				var endDate    = builder.GetExpression(2);
-
-				if (startDate is null || endDate is null)
-				{
-					builder.IsConvertible = false;
-					return;
-				}
-
-				var secondsExpr = builder.Mul<int>(builder.Sub<int>(
-						new FunctionWord(typeof(int), "Days", endDate),
-						new FunctionWord(typeof(int), "Days", startDate)),
-					new ValueWord(86400));
-
-				var midnight = builder.Sub<int>(
-					new FunctionWord(typeof(int), "MIDNIGHT_SECONDS", endDate),
-					new FunctionWord(typeof(int), "MIDNIGHT_SECONDS", startDate));
-
-				var resultExpr = builder.Add<int>(secondsExpr, midnight);
-
-				switch (part)
-				{
-					case DateParts.Day         : resultExpr = builder.Div(resultExpr, 86400); break;
-					case DateParts.Hour        : resultExpr = builder.Div(resultExpr, 3600);  break;
-					case DateParts.Minute      : resultExpr = builder.Div(resultExpr, 60);    break;
-					case DateParts.Second      : break;
-					case DateParts.Millisecond :
-						resultExpr = builder.Add<int>(
-							builder.Mul(resultExpr, 1000),
-							builder.Div(
-								builder.Sub<int>(
-									new FunctionWord(typeof(int), "MICROSECOND", endDate),
-									new FunctionWord(typeof(int), "MICROSECOND", startDate)),
-								1000));
-						break;
-					default:
-						throw new InvalidOperationException($"Unexpected datepart: {part}");
-				}
-
-				builder.ResultExpression = resultExpr;
-			}
-		}
-
-		sealed class DateDiffBuilderClickHouse : IExtensionCallBuilder
-		{
-			public void Build(ISqExtensionBuilder builder)
-			{
-				var part       = builder.GetValue<DateParts>(0);
-				var startDate  = builder.GetExpression(1);
-				var endDate    = builder.GetExpression(2);
-
-				if (startDate is null || endDate is null)
-				{
-					builder.IsConvertible = false;
-					return;
-				}
-
-				string? unit = null;
-				switch (part)
-				{
-					case DateParts.Year   : unit = "year"   ; break;
-					case DateParts.Quarter: unit = "quarter"; break;
-					case DateParts.Month  : unit = "month"  ; break;
-					case DateParts.Week   : unit = "week"   ; break;
-					case DateParts.Day    : unit = "day"    ; break;
-					case DateParts.Hour   : unit = "hour"   ; break;
-					case DateParts.Minute : unit = "minute" ; break;
-					case DateParts.Second : unit = "second" ; break;
-
-					case DateParts.Millisecond:
-						builder.ResultExpression = new ExpressionWord(
-							typeof(long?),
-							"toUnixTimestamp64Milli(toDateTime64({1}, 3)) - toUnixTimestamp64Milli(toDateTime64({0}, 3))",
-							PrecedenceLv.Subtraction,
-							startDate,
-							endDate);
-						break;
-
-					default:
-						throw new InvalidOperationException($"Unexpected datepart: {part}");
-				}
-
-				if (unit != null)
-					builder.ResultExpression = new FunctionWord(typeof(int), "date_diff", new ValueWord(unit), startDate, endDate);
-			}
-		}
-
 		[CLSCompliant(false)]
 		[Extension(               "",              PreferServerSide = true)]
 		[Extension(PN.MySql,      "",              PreferServerSide = true)]
-		[Extension(PN.DB2,        "",              BuilderType = typeof(DateDiffBuilderDB2))]
-		[Extension(PN.SapHana,    "",              BuilderType = typeof(DateDiffBuilderSapHana))]
+		[Extension(PN.DB2,        "",              PreferServerSide = true)]
+		[Extension(PN.SapHana,    "",              PreferServerSide = true)]
 		[Extension(PN.SQLite,     "",              PreferServerSide = true)]
 		[Extension(PN.Oracle,     "",              PreferServerSide = true)]
 		[Extension(PN.PostgreSQL, "",              PreferServerSide = true)]
 		[Extension(PN.Access,     "",              PreferServerSide = true)]
-		[Extension(PN.ClickHouse, "",              BuilderType = typeof(DateDiffBuilderClickHouse))]
+		[Extension(PN.ClickHouse, "",              PreferServerSide = true)]
 		public static int? DateDiff(DateParts part, DateTime? startDate, DateTime? endDate)
 		{
 			if (startDate == null || endDate == null)